NAM-1200 管理アプリ：データmodel変換処理
* 車載線量計バリデータクラス実装
* クラスターモデル、ログモデルにtoString()実装
* 放射線アップロード時のmodel変換処理実装
* 測定種別保存処理追加

担当：伊藤

+++ www_dev/app/modules/view/AbstractView.js
@@ -152,7 +152,7 @@ define(function(require, exports, module) {
                 return;
             }
             var increaseValue = value || this.perProgress;
-            var currentValue = parseInt(this.$progressBar.attr("value"));
+            var currentValue = parseFloat(this.$progressBar.attr("value"));
             this.$progressBar.attr("value",currentValue + increaseValue);
         },
         /**
